Move trim_http_whitespace to net_traits.
