+++ tests/unit/net_traits/lib.rs
@@ -5,3 +5,25 @@
 extern crate net_traits;
 
 #[cfg(test)] mod image;
+
+#[test]
+fn test_trim_http_whitespace() {
+    fn test_trim(in_: &[u8], out: &[u8]) {
+        let b = net_traits::trim_http_whitespace(in_);
+        assert_eq!(b, out);
+    }
+
+    test_trim(b"", b"");
+
+    test_trim(b" ", b"");
+    test_trim(b"a", b"a");
+    test_trim(b" a", b"a");
+    test_trim(b"a ", b"a");
+    test_trim(b" a ", b"a");
+
+    test_trim(b"\t", b"");
+    test_trim(b"a", b"a");
+    test_trim(b"\ta", b"a");
+    test_trim(b"a\t", b"a");
+    test_trim(b"\ta\t", b"a");
+}
